Create Company routes

+++ src/companies/companies.controller.ts
@@ -1,9 +1,11 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, ForbiddenException } from '@nestjs/common';
 import { ApiBearerAuth, ApiInternalServerErrorResponse, ApiOkResponse, ApiOperation, ApiTags, ApiUnauthorizedResponse } from '@nestjs/swagger';
 import { Action } from 'src/auth/enums';
 import { AppAbility } from 'src/casl/casl-ability.factory';
 import { CheckPolicies } from 'src/casl/check-policy.decorator';
 import { PoliciesGuard } from 'src/casl/PoliciesGuard';
+import { User } from '../users/models/user.model';
+import { getUser } from '../users/user.decorator';
 import { CompaniesService } from './companies.service';
 import { CreateCompanyDto } from './dto/create-company.dto';
 import { ReadCompanyDto } from './dto/read-company-dto';
@@ -31,8 +33,8 @@ export class CompaniesController {
   @Get()
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => ability.can(Action.Read, 'Company'))
-  findAll() {
-    return this.companiesService.findAll();
+  findAll(@getUser() auth: User) {
+    return this.companiesService.findAll(!auth.isAdmin ? auth.id : null);
   }
 
   @ApiOperation({summary: "Get a company", description: "Get specific company associated with the logged-in user"})
@@ -42,8 +44,13 @@ export class CompaniesController {
   @Get(':id')
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => ability.can(Action.Read, 'Company'))
-  findOne(@Param('id') id: string) {
-    return this.companiesService.findOne(+id);
+  async findOne(@Param('id') id: string, @getUser() auth: User) {
+    const company = await this.companiesService.findOne(+id);
+
+    if(company.user.id != auth.id && !auth.isAdmin)
+      throw new ForbiddenException("You don't have access to this company.");
+
+    return company;
   }
 
   @ApiOperation({summary: "Update a company", description: "Update a specific company associated with the logged-in user"})
@@ -53,7 +60,12 @@ export class CompaniesController {
   @Patch(':id')
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => ability.can(Action.Update, 'Company'))
-  update(@Param('id') id: string, @Body() updateCompanyDto: UpdateCompanyDto) {
+  async update(@Param('id') id: string, @Body() updateCompanyDto: UpdateCompanyDto, @getUser() auth: User) {
+    const company = await this.findOne(id, auth);
+
+    if(company.user.id != auth.id && !auth.isAdmin)
+      throw new ForbiddenException("You don't have access to this company.");
+
     return this.companiesService.update(+id, updateCompanyDto);
   }
 
@@ -64,7 +76,12 @@ export class CompaniesController {
   @Delete(':id')
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => ability.can(Action.Delete, 'Company'))
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @getUser() auth: User) {
+    const company = await this.findOne(id, auth);
+
+    if(company.user.id != auth.id && !auth.isAdmin)
+      throw new ForbiddenException("You don't have access to this company.");
+
     return this.companiesService.remove(+id);
   }
 }
